Remove storage permission

+++ app/src/main/java/org/meowcat/edxposed/manager/XposedApp.java
@@ -105,31 +105,35 @@ public class XposedApp extends de.robv.android.xposed.installer.XposedApp implem
 
     public void onCreate() {
         super.onCreate();
-        try {
-            Thread.setDefaultUncaughtExceptionHandler((thread, throwable) -> {
+        if (!BuildConfig.DEBUG) {
+            try {
+                Thread.setDefaultUncaughtExceptionHandler((thread, throwable) -> {
 
-                StringWriter sw = new StringWriter();
-                PrintWriter pw = new PrintWriter(sw);
-                throwable.printStackTrace(pw);
-                String stackTraceString = sw.toString();
+                    StringWriter sw = new StringWriter();
+                    PrintWriter pw = new PrintWriter(sw);
+                    throwable.printStackTrace(pw);
+                    String stackTraceString = sw.toString();
 
-                //Reduce data to 128KB so we don't get a TransactionTooLargeException when sending the intent.
-                //The limit is 1MB on Android but some devices seem to have it lower.
-                //See: http://developer.android.com/reference/android/os/TransactionTooLargeException.html
-                //And: http://stackoverflow.com/questions/11451393/what-to-do-on-transactiontoolargeexception#comment46697371_12809171
-                if (stackTraceString.length() > 131071) {
-                    String disclaimer = " [stack trace too large]";
-                    stackTraceString = stackTraceString.substring(0, 131071 - disclaimer.length()) + disclaimer;
-                }
-                Intent intent = new Intent(XposedApp.this, CrashReportActivity.class);
-                intent.putExtra(BuildConfig.APPLICATION_ID + ".EXTRA_STACK_TRACE", stackTraceString);
-                intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-                XposedApp.this.startActivity(intent);
-                android.os.Process.killProcess(android.os.Process.myPid());
-                System.exit(10);
-            });
-        } catch (Throwable t) {
+                    //Reduce data to 128KB so we don't get a TransactionTooLargeException when sending the intent.
+                    //The limit is 1MB on Android but some devices seem to have it lower.
+                    //See: http://developer.android.com/reference/android/os/TransactionTooLargeException.html
+                    //And: http://stackoverflow.com/questions/11451393/what-to-do-on-transactiontoolargeexception#comment46697371_12809171
+                    if (stackTraceString.length() > 131071) {
+                        String disclaimer = " [stack trace too large]";
+                        stackTraceString = stackTraceString.substring(0, 131071 - disclaimer.length()) + disclaimer;
+                    }
+                    Intent intent = new Intent(XposedApp.this, CrashReportActivity.class);
+                    intent.putExtra(BuildConfig.APPLICATION_ID + ".EXTRA_STACK_TRACE", stackTraceString);
+                    intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+                    XposedApp.this.startActivity(intent);
+                    android.os.Process.killProcess(android.os.Process.myPid());
+                    System.exit(10);
+                });
+            } catch (Throwable t) {
+                t.printStackTrace();
+            }
         }
+
         mInstance = this;
         mUiThread = Thread.currentThread();
         mMainHandler = new Handler();
